@@ -3,22 +3,22 @@ package com.hexagram2021.subject3.common;
 import com.hexagram2021.subject3.Subject3;
 import com.hexagram2021.subject3.common.entities.IBedVehicle;
 import com.hexagram2021.subject3.common.entities.IHasVehicleRespawnPosition;
-import net.minecraft.block.PortalInfo;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.Util;
-import net.minecraft.util.math.ChunkPos;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraft.world.World;
-import net.minecraft.world.server.ServerWorld;
+import net.minecraft.Util;
+import net.minecraft.core.SectionPos;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.ChunkPos;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.portal.PortalInfo;
+import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.common.util.ITeleporter;
 import net.minecraftforge.event.entity.EntityEvent;
 import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
-import javax.annotation.Nullable;
 import java.util.UUID;
 import java.util.function.Function;
 
@@ -28,25 +28,25 @@ import static com.hexagram2021.subject3.Subject3.MODID;
 public class STEventHandler {
 	@SubscribeEvent
 	public static void onPlayerRespawn(PlayerEvent.PlayerRespawnEvent event) {
-		PlayerEntity player = event.getPlayer();
+		Player player = event.getPlayer();
 		if(player instanceof IHasVehicleRespawnPosition) {
 			UUID bedVehicleUUID = ((IHasVehicleRespawnPosition) player).getBedVehicleUUID();
 			if(bedVehicleUUID != null) {
-				Entity bedVehicle = ((ServerWorld) player.level).getEntity(bedVehicleUUID);
+				Entity bedVehicle = ((ServerLevel) player.level).getEntity(bedVehicleUUID);
 				if (bedVehicle instanceof IBedVehicle) {
 					if (((IBedVehicle) bedVehicle).passengersCount() == 0) {
 						if(player.level == bedVehicle.level) {
 							player.setPos(bedVehicle.getX(), bedVehicle.getY(), bedVehicle.getZ());
 						} else {
-							player.changeDimension((ServerWorld) player.level, new ITeleporter() {
+							player.changeDimension((ServerLevel) player.level, new ITeleporter() {
 								@Override
-								public Entity placeEntity(Entity entity, ServerWorld currentWorld, ServerWorld destWorld, float yaw, Function<Boolean, Entity> repositionEntity) {
+								public Entity placeEntity(Entity entity, ServerLevel currentWorld, ServerLevel destWorld, float yaw, Function<Boolean, Entity> repositionEntity) {
 									return repositionEntity.apply(false);
 								}
 
 								@Override
-								public PortalInfo getPortalInfo(Entity entity, ServerWorld destWorld, Function<ServerWorld, PortalInfo> defaultPortalInfo) {
-									return new PortalInfo(bedVehicle.position(), Vector3d.ZERO, entity.yRot, entity.xRot);
+								public PortalInfo getPortalInfo(Entity entity, ServerLevel destWorld, Function<ServerLevel, PortalInfo> defaultPortalInfo) {
+									return new PortalInfo(bedVehicle.position(), Vec3.ZERO, entity.getYRot(), entity.getXRot());
 								}
 							});
 						}
@@ -56,28 +56,28 @@ public class STEventHandler {
 				} else {
 					((IHasVehicleRespawnPosition) player).setBedVehicleUUID(null);
 				}
-				player.sendMessage(new TranslationTextComponent("message.subject3.bed_vehicle_occupied"), Util.NIL_UUID);
+				player.sendMessage(new TranslatableComponent("message.subject3.bed_vehicle_occupied"), Util.NIL_UUID);
 			}
 		}
 	}
 
 	@SubscribeEvent
 	public static void onPlayerClone(PlayerEvent.Clone event) {
-		PlayerEntity player = event.getPlayer();
-		PlayerEntity original = event.getOriginal();
+		Player player = event.getPlayer();
+		Player original = event.getOriginal();
 		if(player instanceof IHasVehicleRespawnPosition && original instanceof IHasVehicleRespawnPosition) {
 			((IHasVehicleRespawnPosition)player).setBedVehicleUUID(((IHasVehicleRespawnPosition)original).getBedVehicleUUID());
 		}
 	}
 
 	@SubscribeEvent
-	public static void onEntityEnterChunk(EntityEvent.EnteringChunk event) {
+	public static void onEntityEnterChunk(EntityEvent.EnteringSection event) {
 		if(event.getEntity() instanceof IBedVehicle) {
-			World level = event.getEntity().level;
-			if(level instanceof ServerWorld) {
-				ServerWorld serverlevel = (ServerWorld) level;
-				if (level.dimension().equals(ServerWorld.OVERWORLD)) {
-					ChunkPos newPos = new ChunkPos(event.getNewChunkX(), event.getNewChunkZ());
+			Level level = event.getEntity().level;
+			if(level instanceof ServerLevel serverlevel) {
+				if (level.dimension().equals(ServerLevel.OVERWORLD)) {
+					SectionPos sectionPos = event.getNewPos();
+					ChunkPos newPos = new ChunkPos(sectionPos.getX(), sectionPos.getZ());
 
 					ChunkPos oldPos = STSavedData.addBedVehicle(event.getEntity().getUUID(), newPos);
 
