@@ -4,43 +4,42 @@ import com.hexagram2021.subject3.common.STSavedData;
 import com.hexagram2021.subject3.register.STBlocks;
 import com.hexagram2021.subject3.register.STEntities;
 import com.hexagram2021.subject3.register.STItems;
-import net.minecraft.block.BedBlock;
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.EntityType;
-import net.minecraft.entity.item.minecart.AbstractMinecartEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.DyeColor;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.network.IPacket;
-import net.minecraft.network.datasync.DataParameter;
-import net.minecraft.network.datasync.DataSerializers;
-import net.minecraft.network.datasync.EntityDataManager;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.DamageSource;
-import net.minecraft.util.Hand;
-import net.minecraft.util.math.ChunkPos;
-import net.minecraft.world.Explosion;
-import net.minecraft.world.GameRules;
-import net.minecraft.world.World;
-import net.minecraft.world.server.ServerWorld;
-import net.minecraftforge.common.util.Constants;
-import net.minecraftforge.fml.network.NetworkHooks;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.Tag;
+import net.minecraft.network.protocol.Packet;
+import net.minecraft.network.syncher.EntityDataAccessor;
+import net.minecraft.network.syncher.EntityDataSerializers;
+import net.minecraft.network.syncher.SynchedEntityData;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.vehicle.AbstractMinecart;
+import net.minecraft.world.item.DyeColor;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.level.ChunkPos;
+import net.minecraft.world.level.Explosion;
+import net.minecraft.world.level.GameRules;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.BedBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraftforge.fmllegacy.network.NetworkHooks;
 
 import javax.annotation.Nonnull;
 
-public class BedMinecartEntity extends AbstractMinecartEntity implements IBedVehicle {
+public class BedMinecartEntity extends AbstractMinecart implements IBedVehicle {
 	public static Type BED;
 
-	private static final DataParameter<Integer> DATA_ID_DYE_COLOR = EntityDataManager.defineId(BedMinecartEntity.class, DataSerializers.INT);
+	private static final EntityDataAccessor<Integer> DATA_ID_DYE_COLOR = SynchedEntityData.defineId(BedMinecartEntity.class, EntityDataSerializers.INT);
 
-	public BedMinecartEntity(EntityType<?> type, World level) {
+	public BedMinecartEntity(EntityType<?> type, Level level) {
 		super(type, level);
-		this.forcedLoading = true;
 	}
 
-	public BedMinecartEntity(World level, double x, double y, double z) {
+	public BedMinecartEntity(Level level, double x, double y, double z) {
 		super(STEntities.BED_MINECART.get(), level, x, y, z);
 	}
 
@@ -51,38 +50,41 @@ public class BedMinecartEntity extends AbstractMinecartEntity implements IBedVeh
 	}
 
 	@Override @Nonnull
-	public ActionResultType interact(@Nonnull PlayerEntity player, @Nonnull Hand hand) {
-		ActionResultType ret = super.interact(player, hand);
+	public InteractionResult interact(@Nonnull Player player, @Nonnull InteractionHand hand) {
+		InteractionResult ret = super.interact(player, hand);
 		if (ret.consumesAction()) return ret;
 		if (player.isSecondaryUseActive()) {
-			return ActionResultType.PASS;
+			return InteractionResult.PASS;
 		}
 		if (this.isVehicle()) {
-			return ActionResultType.PASS;
+			return InteractionResult.PASS;
 		}
 		if (!this.level.isClientSide) {
 			if (!BedBlock.canSetSpawn(this.level)) {
-				this.level.explode(this, DamageSource.badRespawnPointExplosion(), null, this.getX() + 0.5D, this.getY() + 0.125D, this.getZ() + 0.5D, 5.0F, true, Explosion.Mode.DESTROY);
+				this.level.explode(
+						this, DamageSource.badRespawnPointExplosion(), null,
+						this.getX() + 0.5D, this.getY() + 0.125D, this.getZ() + 0.5D,
+						5.0F, true, Explosion.BlockInteraction.DESTROY
+				);
 				this.removeBedVehicle();
-				return ActionResultType.SUCCESS;
+				return InteractionResult.SUCCESS;
 			}
 			if(player.startRiding(this)) {
-				ret = ActionResultType.CONSUME;
+				ret = InteractionResult.CONSUME;
 				if(player instanceof IHasVehicleRespawnPosition) {
 					((IHasVehicleRespawnPosition)player).setBedVehicleUUID(this.uuid);
 				}
 			} else {
-				ret = ActionResultType.PASS;
+				ret = InteractionResult.PASS;
 			}
 			return ret;
 		}
-		return ActionResultType.SUCCESS;
+		return InteractionResult.SUCCESS;
 	}
 
-	@SuppressWarnings("deprecation")
 	@Override
 	public boolean hurt(@Nonnull DamageSource damageSource, float value) {
-		if (!this.level.isClientSide && !this.removed) {
+		if (!this.level.isClientSide && !this.isRemoved()) {
 			if (this.isInvulnerableTo(damageSource)) {
 				return false;
 			}
@@ -90,7 +92,7 @@ public class BedMinecartEntity extends AbstractMinecartEntity implements IBedVeh
 			this.setHurtTime(10);
 			this.markHurt();
 			this.setDamage(this.getDamage() + value * 10.0F);
-			boolean flag = damageSource.getEntity() instanceof PlayerEntity && ((PlayerEntity)damageSource.getEntity()).abilities.instabuild;
+			boolean flag = damageSource.getEntity() instanceof Player player && player.getAbilities().instabuild;
 			if (flag || this.getDamage() > 40.0F) {
 				this.ejectPassengers();
 				if (flag && !this.hasCustomName()) {
@@ -143,15 +145,15 @@ public class BedMinecartEntity extends AbstractMinecartEntity implements IBedVeh
 	}
 
 	@Override
-	protected void addAdditionalSaveData(@Nonnull CompoundNBT nbt) {
+	protected void addAdditionalSaveData(@Nonnull CompoundTag nbt) {
 		super.addAdditionalSaveData(nbt);
 		nbt.putString("DyeColor", this.getBedColor().getName());
 	}
 
 	@Override
-	protected void readAdditionalSaveData(@Nonnull CompoundNBT nbt) {
+	protected void readAdditionalSaveData(@Nonnull CompoundTag nbt) {
 		super.readAdditionalSaveData(nbt);
-		if (nbt.contains("DyeColor", Constants.NBT.TAG_STRING)) {
+		if (nbt.contains("DyeColor", Tag.TAG_STRING)) {
 			this.setColor(DyeColor.byName(nbt.getString("DyeColor"), DyeColor.WHITE));
 		}
 	}
@@ -162,22 +164,22 @@ public class BedMinecartEntity extends AbstractMinecartEntity implements IBedVeh
 	}
 
 	@Override @Nonnull
-	public IPacket<?> getAddEntityPacket() {
+	public Packet<?> getAddEntityPacket() {
 		return NetworkHooks.getEntitySpawningPacket(this);
 	}
 
 	@Override
 	public void removeBedVehicle() {
 		ChunkPos chunkPos = STSavedData.removeBedVehicle(this.uuid);
-		if(chunkPos != null && this.level instanceof ServerWorld) {
-			STSavedData.updateForceChunk(chunkPos, (ServerWorld)this.level, false);
+		if(chunkPos != null && this.level instanceof ServerLevel serverLevel) {
+			STSavedData.updateForceChunk(chunkPos, serverLevel, false);
 		}
-		this.remove();
+		this.kill();
 	}
 
 	@Override
 	public float getBedVehicleRotY() {
-		return this.yRot + 90.0f;
+		return this.getYRot() + 90.0f;
 	}
 	@Override
 	public double getBedVehicleOffsetY() {
@@ -194,7 +196,7 @@ public class BedMinecartEntity extends AbstractMinecartEntity implements IBedVeh
 		return false;
 	}
 
-	public static BedMinecartEntity createBedMinecart(World level, double x, double y, double z) {
+	public static BedMinecartEntity createBedMinecart(Level level, double x, double y, double z) {
 		return new BedMinecartEntity(level, x, y, z);
 	}
 }
