@@ -1,14 +1,16 @@
 package com.hexagram2021.subject3.register;
 
-import net.minecraft.block.*;
-import net.minecraft.block.material.Material;
-import net.minecraft.item.DyeColor;
-import net.minecraft.item.Item;
-import net.minecraft.util.IItemProvider;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.Util;
+import net.minecraft.Util;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.DyeColor;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.level.ItemLike;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.material.Material;
 import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.fml.RegistryObject;
+import net.minecraftforge.fmllegacy.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 
@@ -23,8 +25,8 @@ public class STBlocks {
 	public static final DeferredRegister<Block> REGISTER = DeferredRegister.create(ForgeRegistries.BLOCKS, MODID);
 
 	public static class Technical {
-		private static final Supplier<AbstractBlock.Properties> PROPERTIES = () ->
-				AbstractBlock.Properties.of(Material.STRUCTURAL_AIR).noCollission().noDrops().instabreak();
+		private static final Supplier<BlockBehaviour.Properties> PROPERTIES = () ->
+				BlockBehaviour.Properties.of(Material.STRUCTURAL_AIR).noCollission().noDrops().instabreak();
 
 		public static final BlockEntry<Block> BLACK_BOAT_BED = new BlockEntry<>("black_boat_bed", PROPERTIES, Block::new);
 		public static final BlockEntry<Block> BLUE_BOAT_BED = new BlockEntry<>("blue_boat_bed", PROPERTIES, Block::new);
@@ -61,79 +63,45 @@ public class STBlocks {
 		public static final BlockEntry<Block> YELLOW_MINECART_BED = new BlockEntry<>("yellow_minecart_bed", PROPERTIES, Block::new);
 
 		public static BlockState getBoatBedBlockState(DyeColor color) {
-			switch (color) {
-				case BLACK:
-					return Technical.BLACK_BOAT_BED.defaultBlockState();
-				case BLUE:
-					return Technical.BLUE_BOAT_BED.defaultBlockState();
-				case BROWN:
-					return Technical.BROWN_BOAT_BED.defaultBlockState();
-				case CYAN:
-					return Technical.CYAN_BOAT_BED.defaultBlockState();
-				case GRAY:
-					return Technical.GRAY_BOAT_BED.defaultBlockState();
-				case GREEN:
-					return Technical.GREEN_BOAT_BED.defaultBlockState();
-				case LIGHT_BLUE:
-					return Technical.LIGHT_BLUE_BOAT_BED.defaultBlockState();
-				case LIGHT_GRAY:
-					return Technical.LIGHT_GRAY_BOAT_BED.defaultBlockState();
-				case LIME:
-					return Technical.LIME_BOAT_BED.defaultBlockState();
-				case MAGENTA:
-					return Technical.MAGENTA_BOAT_BED.defaultBlockState();
-				case ORANGE:
-					return Technical.ORANGE_BOAT_BED.defaultBlockState();
-				case PINK:
-					return Technical.PINK_BOAT_BED.defaultBlockState();
-				case PURPLE:
-					return Technical.PURPLE_BOAT_BED.defaultBlockState();
-				case RED:
-					return Technical.RED_BOAT_BED.defaultBlockState();
-				case WHITE:
-				default:
-					return Technical.WHITE_BOAT_BED.defaultBlockState();
-				case YELLOW:
-					return Technical.YELLOW_BOAT_BED.defaultBlockState();
-			}
+			return switch (color) {
+				case BLACK -> Technical.BLACK_BOAT_BED.defaultBlockState();
+				case BLUE -> Technical.BLUE_BOAT_BED.defaultBlockState();
+				case BROWN -> Technical.BROWN_BOAT_BED.defaultBlockState();
+				case CYAN -> Technical.CYAN_BOAT_BED.defaultBlockState();
+				case GRAY -> Technical.GRAY_BOAT_BED.defaultBlockState();
+				case GREEN -> Technical.GREEN_BOAT_BED.defaultBlockState();
+				case LIGHT_BLUE -> Technical.LIGHT_BLUE_BOAT_BED.defaultBlockState();
+				case LIGHT_GRAY -> Technical.LIGHT_GRAY_BOAT_BED.defaultBlockState();
+				case LIME -> Technical.LIME_BOAT_BED.defaultBlockState();
+				case MAGENTA -> Technical.MAGENTA_BOAT_BED.defaultBlockState();
+				case ORANGE -> Technical.ORANGE_BOAT_BED.defaultBlockState();
+				case PINK -> Technical.PINK_BOAT_BED.defaultBlockState();
+				case PURPLE -> Technical.PURPLE_BOAT_BED.defaultBlockState();
+				case RED -> Technical.RED_BOAT_BED.defaultBlockState();
+				case WHITE -> Technical.WHITE_BOAT_BED.defaultBlockState();
+				case YELLOW -> Technical.YELLOW_BOAT_BED.defaultBlockState();
+			};
 		}
 
 		public static BlockState getMinecartBedBlockState(DyeColor color) {
-			switch (color) {
-				case BLACK:
-					return Technical.BLACK_MINECART_BED.defaultBlockState();
-				case BLUE:
-					return Technical.BLUE_MINECART_BED.defaultBlockState();
-				case BROWN:
-					return Technical.BROWN_MINECART_BED.defaultBlockState();
-				case CYAN:
-					return Technical.CYAN_MINECART_BED.defaultBlockState();
-				case GRAY:
-					return Technical.GRAY_MINECART_BED.defaultBlockState();
-				case GREEN:
-					return Technical.GREEN_MINECART_BED.defaultBlockState();
-				case LIGHT_BLUE:
-					return Technical.LIGHT_BLUE_MINECART_BED.defaultBlockState();
-				case LIGHT_GRAY:
-					return Technical.LIGHT_GRAY_MINECART_BED.defaultBlockState();
-				case LIME:
-					return Technical.LIME_MINECART_BED.defaultBlockState();
-				case MAGENTA:
-					return Technical.MAGENTA_MINECART_BED.defaultBlockState();
-				case ORANGE:
-					return Technical.ORANGE_MINECART_BED.defaultBlockState();
-				case PINK:
-					return Technical.PINK_MINECART_BED.defaultBlockState();
-				case PURPLE:
-					return Technical.PURPLE_MINECART_BED.defaultBlockState();
-				case RED:
-					return Technical.RED_MINECART_BED.defaultBlockState();
-				case WHITE:
-				default:
-					return Technical.WHITE_MINECART_BED.defaultBlockState();
-				case YELLOW:
-					return Technical.YELLOW_MINECART_BED.defaultBlockState();
-			}
+			return switch (color) {
+				case BLACK -> Technical.BLACK_MINECART_BED.defaultBlockState();
+				case BLUE -> Technical.BLUE_MINECART_BED.defaultBlockState();
+				case BROWN -> Technical.BROWN_MINECART_BED.defaultBlockState();
+				case CYAN -> Technical.CYAN_MINECART_BED.defaultBlockState();
+				case GRAY -> Technical.GRAY_MINECART_BED.defaultBlockState();
+				case GREEN -> Technical.GREEN_MINECART_BED.defaultBlockState();
+				case LIGHT_BLUE -> Technical.LIGHT_BLUE_MINECART_BED.defaultBlockState();
+				case LIGHT_GRAY -> Technical.LIGHT_GRAY_MINECART_BED.defaultBlockState();
+				case LIME -> Technical.LIME_MINECART_BED.defaultBlockState();
+				case MAGENTA -> Technical.MAGENTA_MINECART_BED.defaultBlockState();
+				case ORANGE -> Technical.ORANGE_MINECART_BED.defaultBlockState();
+				case PINK -> Technical.PINK_MINECART_BED.defaultBlockState();
+				case PURPLE -> Technical.PURPLE_MINECART_BED.defaultBlockState();
+				case RED -> Technical.RED_MINECART_BED.defaultBlockState();
+				case WHITE -> Technical.WHITE_MINECART_BED.defaultBlockState();
+				case YELLOW -> Technical.YELLOW_MINECART_BED.defaultBlockState();
+			};
 		}
 
 		public static void init() {}
@@ -146,15 +114,15 @@ public class STBlocks {
 	}
 
 	@SuppressWarnings("unused")
-	public static final class BlockEntry<T extends Block> implements Supplier<T>, IItemProvider {
+	public static final class BlockEntry<T extends Block> implements Supplier<T>, ItemLike {
 		private final RegistryObject<T> regObject;
-		private final Supplier<AbstractBlock.Properties> properties;
+		private final Supplier<BlockBehaviour.Properties> properties;
 
-		public static BlockEntry<Block> simple(String name, Supplier<AbstractBlock.Properties> properties, Consumer<Block> extra) {
+		public static BlockEntry<Block> simple(String name, Supplier<BlockBehaviour.Properties> properties, Consumer<Block> extra) {
 			return new BlockEntry<>(name, properties, p -> Util.make(new Block(p), extra));
 		}
 
-		public BlockEntry(String name, Supplier<AbstractBlock.Properties> properties, Function<AbstractBlock.Properties, T> make) {
+		public BlockEntry(String name, Supplier<BlockBehaviour.Properties> properties, Function<BlockBehaviour.Properties, T> make) {
 			this.properties = properties;
 			this.regObject = REGISTER.register(name, () -> make.apply(properties.get()));
 		}
@@ -173,7 +141,7 @@ public class STBlocks {
 			return regObject.getId();
 		}
 
-		public AbstractBlock.Properties getProperties()
+		public BlockBehaviour.Properties getProperties()
 		{
 			return properties.get();
 		}
