@@ -1,45 +1,56 @@
 package com.hexagram2021.subject3.client.renderers;
 
+import com.google.common.collect.ImmutableMap;
 import com.hexagram2021.subject3.client.layers.BoatBedLayer;
 import com.hexagram2021.subject3.client.models.BedBoatModel;
 import com.hexagram2021.subject3.common.entities.BedBoatEntity;
-import com.mojang.blaze3d.matrix.MatrixStack;
-import com.mojang.blaze3d.vertex.IVertexBuilder;
-import net.minecraft.client.renderer.IRenderTypeBuffer;
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import com.mojang.datafixers.util.Pair;
+import com.mojang.math.Quaternion;
+import com.mojang.math.Vector3f;
+import net.minecraft.client.model.geom.ModelLayerLocation;
+import net.minecraft.client.model.geom.ModelLayers;
+import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.EntityRenderer;
-import net.minecraft.client.renderer.entity.EntityRendererManager;
-import net.minecraft.client.renderer.entity.IEntityRenderer;
-import net.minecraft.client.renderer.entity.layers.LayerRenderer;
+import net.minecraft.client.renderer.entity.EntityRendererProvider;
+import net.minecraft.client.renderer.entity.RenderLayerParent;
+import net.minecraft.client.renderer.entity.layers.RenderLayer;
 import net.minecraft.client.renderer.texture.OverlayTexture;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.MathHelper;
-import net.minecraft.util.math.vector.Quaternion;
-import net.minecraft.util.math.vector.Vector3f;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.Mth;
+import net.minecraft.world.entity.vehicle.Boat;
 
 import javax.annotation.Nonnull;
+import java.util.Map;
+import java.util.stream.Stream;
 
-public class BedBoatRenderer extends EntityRenderer<BedBoatEntity> implements IEntityRenderer<BedBoatEntity, BedBoatModel> {
-	private final LayerRenderer<BedBoatEntity, BedBoatModel> bedLayer;
+import static com.hexagram2021.subject3.Subject3.MODID;
 
-	private static final ResourceLocation[] BOAT_TEXTURE_LOCATIONS = new ResourceLocation[]{
-			new ResourceLocation("textures/entity/boat/oak.png"),
-			new ResourceLocation("textures/entity/boat/spruce.png"),
-			new ResourceLocation("textures/entity/boat/birch.png"),
-			new ResourceLocation("textures/entity/boat/jungle.png"),
-			new ResourceLocation("textures/entity/boat/acacia.png"),
-			new ResourceLocation("textures/entity/boat/dark_oak.png")
-	};
-	protected final BedBoatModel model = new BedBoatModel();
+public class BedBoatRenderer extends EntityRenderer<BedBoatEntity> implements RenderLayerParent<BedBoatEntity, BedBoatModel> {
+	public static final ModelLayerLocation LAYER_LOCATION = new ModelLayerLocation(new ResourceLocation(MODID, "bed_boat"), "main");
 
-	public BedBoatRenderer(EntityRendererManager manager) {
+	private final RenderLayer<BedBoatEntity, BedBoatModel> bedLayer;
+	private final Map<Boat.Type, Pair<ResourceLocation, BedBoatModel>> boatResources;
+	private final BedBoatModel defaultModel;
+
+	public BedBoatRenderer(EntityRendererProvider.Context manager) {
 		super(manager);
 		this.shadowRadius = 0.8F;
 		this.bedLayer = new BoatBedLayer(this);
+		this.boatResources = Stream.of(Boat.Type.values()).collect(ImmutableMap.toImmutableMap(
+				type -> type,
+				type -> Pair.of(
+						new ResourceLocation("textures/entity/boat/" + type.getName() + ".png"),
+						new BedBoatModel(manager.bakeLayer(ModelLayers.createBoatModelName(type)))
+				)
+		));
+		this.defaultModel = new BedBoatModel(manager.bakeLayer(LAYER_LOCATION));
 	}
 
 	@Override
-	public void render(BedBoatEntity bedBoatEntity, float y, float ticks, MatrixStack transform, @Nonnull IRenderTypeBuffer buffer, int h) {
+	public void render(BedBoatEntity bedBoatEntity, float y, float ticks, PoseStack transform, @Nonnull MultiBufferSource buffer, int h) {
 		transform.pushPose();
 		transform.translate(0.0D, 0.375D, 0.0D);
 		transform.mulPose(Vector3f.YP.rotationDegrees(180.0F - y));
@@ -50,37 +61,42 @@ public class BedBoatRenderer extends EntityRenderer<BedBoatEntity> implements IE
 		}
 
 		if (f > 0.0F) {
-			transform.mulPose(Vector3f.XP.rotationDegrees(MathHelper.sin(f) * f * f1 / 10.0F * (float)bedBoatEntity.getHurtDir()));
+			transform.mulPose(Vector3f.XP.rotationDegrees(Mth.sin(f) * f * f1 / 10.0F * (float)bedBoatEntity.getHurtDir()));
 		}
 
 		float f2 = bedBoatEntity.getBubbleAngle(ticks);
-		if (!MathHelper.equal(f2, 0.0F)) {
+		if (!Mth.equal(f2, 0.0F)) {
 			transform.mulPose(new Quaternion(new Vector3f(1.0F, 0.0F, 1.0F), bedBoatEntity.getBubbleAngle(ticks), true));
 		}
 
+		Pair<ResourceLocation, BedBoatModel> pair = this.getModelWithLocation(bedBoatEntity);
+		ResourceLocation resourcelocation = pair.getFirst();
+		BedBoatModel model = pair.getSecond();
 		transform.scale(-1.0F, -1.0F, 1.0F);
 		transform.mulPose(Vector3f.YP.rotationDegrees(90.0F));
-		this.model.setupAnim(bedBoatEntity, ticks, 0.0F, -0.1F, 0.0F, 0.0F);
-		IVertexBuilder ivertexbuilder = buffer.getBuffer(this.model.renderType(this.getTextureLocation(bedBoatEntity)));
-		this.model.renderToBuffer(transform, ivertexbuilder, h, OverlayTexture.NO_OVERLAY, 1.0F, 1.0F, 1.0F, 1.0F);
+		model.setupAnim(bedBoatEntity, ticks, 0.0F, -0.1F, 0.0F, 0.0F);
+		VertexConsumer vertexConsumer = buffer.getBuffer(model.renderType(resourcelocation));
+		model.renderToBuffer(transform, vertexConsumer, h, OverlayTexture.NO_OVERLAY, 1.0F, 1.0F, 1.0F, 1.0F);
 		if (!bedBoatEntity.isUnderWater()) {
-			IVertexBuilder builder = buffer.getBuffer(RenderType.waterMask());
-			this.model.waterPatch().render(transform, builder, h, OverlayTexture.NO_OVERLAY);
+			VertexConsumer builder = buffer.getBuffer(RenderType.waterMask());
+			model.waterPatch().render(transform, builder, h, OverlayTexture.NO_OVERLAY);
 		}
 
-		this.bedLayer.render(transform, buffer, h, bedBoatEntity, 0.0F, 0.0F, ticks, 0.0F, 0.0F, MathHelper.lerp(ticks, bedBoatEntity.xRotO, bedBoatEntity.xRot));
+		this.bedLayer.render(transform, buffer, h, bedBoatEntity, 0.0F, 0.0F, ticks, 0.0F, 0.0F, Mth.lerp(ticks, bedBoatEntity.xRotO, bedBoatEntity.getXRot()));
 
 		transform.popPose();
 		super.render(bedBoatEntity, y, ticks, transform, buffer, h);
 	}
 
+	public Pair<ResourceLocation, BedBoatModel> getModelWithLocation(@Nonnull BedBoatEntity boat) { return this.boatResources.get(boat.getBoatType()); }
+
 	@Override @Nonnull
-	public BedBoatModel getModel() {
-		return this.model;
+	public ResourceLocation getTextureLocation(@Nonnull BedBoatEntity boat) {
+		return this.getModelWithLocation(boat).getFirst();
 	}
 
 	@Override @Nonnull
-	public ResourceLocation getTextureLocation(BedBoatEntity entity) {
-		return BOAT_TEXTURE_LOCATIONS[entity.getBoatType().ordinal()];
+	public BedBoatModel getModel() {
+		return this.defaultModel;
 	}
 }
