@@ -1,11 +1,11 @@
 package com.hexagram2021.subject3.mixin;
 
 import com.hexagram2021.subject3.common.entities.IHasVehicleRespawnPosition;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.util.Util;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraftforge.common.util.Constants;
+import net.minecraft.Util;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.Tag;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.server.level.ServerPlayer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -15,20 +15,20 @@ import javax.annotation.Nullable;
 import java.util.Objects;
 import java.util.UUID;
 
-@Mixin(ServerPlayerEntity.class)
+@Mixin(ServerPlayer.class)
 public class ServerPlayerEntityMixin implements IHasVehicleRespawnPosition {
 	@Nullable
 	private UUID respawnBedVehicle;
 
 	@Inject(method = "readAdditionalSaveData", at = @At(value = "TAIL"))
-	public void readBedVehicleRespawnUUID(CompoundNBT nbt, CallbackInfo ci) {
-		if(nbt.contains("BedVehicle", Constants.NBT.TAG_INT_ARRAY)) {
+	public void readBedVehicleRespawnUUID(CompoundTag nbt, CallbackInfo ci) {
+		if(nbt.contains("BedVehicle", Tag.TAG_INT_ARRAY)) {
 			this.respawnBedVehicle = nbt.getUUID("BedVehicle");
 		}
 	}
 
 	@Inject(method = "addAdditionalSaveData", at = @At(value = "TAIL"))
-	public void addBedVehicleRespawnUUID(CompoundNBT nbt, CallbackInfo ci) {
+	public void addBedVehicleRespawnUUID(CompoundTag nbt, CallbackInfo ci) {
 		if(this.respawnBedVehicle != null) {
 			nbt.putUUID("BedVehicle", this.respawnBedVehicle);
 		}
@@ -43,7 +43,7 @@ public class ServerPlayerEntityMixin implements IHasVehicleRespawnPosition {
 	public void setBedVehicleUUID(@Nullable UUID uuid) {
 		if(!Objects.equals(uuid, this.respawnBedVehicle)) {
 			this.respawnBedVehicle = uuid;
-			((ServerPlayerEntity)(Object)this).sendMessage(new TranslationTextComponent("block.minecraft.set_spawn"), Util.NIL_UUID);
+			((ServerPlayer)(Object)this).sendMessage(new TranslatableComponent("block.minecraft.set_spawn"), Util.NIL_UUID);
 		}
 	}
 }
